Move SelectFormGroup and TextFormGroup to their own files
